requirements.txt
Flask
Gunicorn
Pillow
Flask-SocketIO